Move Choices to pretalx.common.models
