+++ src/pretalx/submission/models/submission.py
@@ -15,9 +15,9 @@ from django.utils.translation import gettext_lazy as _
 from django.utils.translation import pgettext
 from django_scopes import ScopedManager
 
-from pretalx.common.choices import Choices
 from pretalx.common.exceptions import SubmissionError
 from pretalx.common.mixins.models import GenerateCode, PretalxModel
+from pretalx.common.models.choices import Choices
 from pretalx.common.phrases import phrases
 from pretalx.common.urls import EventUrls
 from pretalx.common.utils import path_with_hash
